Update tests to follow "throws" convention
Standardized tests so that they all throw when they encounter an error. As a result, some of these tests now perform "validation" that they didn't before.

However, in one case (testImageAndLinkMetaData), the test no longer compares the first 10 characters of the exported string to the imported string. This kind of test – comparing exported data to imported data – seems very useful, and it would be ideal if validateDocX() handled that. Then, many tests would benefit.