Write “modern” docx files that don’t show “Compatibility Mode” in Word
This required adding a “settings.xml” file that contains the following important line:
<w:compatSetting w:name="compatibilityMode" w:uri="http://schemas.microsoft.com/office/word" w:val="15"/>

Setting “compatibilityMode” to 15 indicates that this is a “modern” docx file. With that set, Word will not open the docx in “Compatibility Mode” and you will be able to use recent Word features.

In order for the settings.xml file to be found, “document.xml.rels” now contains a reference to “settings.xml.”

Lastly, “[Content_Types].xml” was edited so that it is more human-readable.