Allow clients to force indentation to 0
When translating NSParagraphStyle values to docx, a value of 0 is often used to indicate that no attributes should be output. This is true for paragraph spacing, line height, etc. And, for many of those, a value of 0 doesn’t really make sense.

For indentation, though, setting values to zero can be valid. We don't want to break existing clients (which might rely on this behavior), so we've introduced a constant, `zeroIndent`, which can be used to force indentation to zero, when desired.