Oneeightyg issue17 (#33)
* DocX can now create Word documents with styles

Introduced a new DocXConfiguration class that allows clients to control certain aspects of docx creation:
  o) The styles.xml file to be included in the docx
  o) Whether the declared font should always be output

Added two new NSAttributedString keys so that clients can indicate  that text should use Word paragraph and character styles:
  o) NSAttributedString.Key.paragraphStyleId
  o) NSAttributedString.Key.characterStyleId

The value for these attributes should be a style id that is present in the Word file.

* Allow clients to force indentation to 0

When translating NSParagraphStyle values to docx, a value of 0 is often used to indicate that no attributes should be output. This is true for paragraph spacing, line height, etc. And, for many of those, a value of 0 doesn’t really make sense.

For indentation, though, setting values to zero can be valid. We don't want to break existing clients (which might rely on this behavior), so we've introduced a constant, `zeroIndent`, which can be used to force indentation to zero, when desired.

* Handle baseline offset NSAttributedString.Key

We assume that a baseline offset of less than zero indicates subscript, and a baseline offset of more than zero indicates superscript.

* Revert "Handle baseline offset NSAttributedString.Key"

This reverts commit c790d923910bb90effa541a8b4f3c789f0ac74c5.

Pulling this out of issue17 so that I can file a separate PR for it.

* Revert "Allow clients to force indentation to 0"

This reverts commit f983b7ad28e209b341b37b77846dc1809043533f.

Pulling this out of issue17 so that I can file a separate PR for it.

* DocXConfiguration -> DocXStyleConfiguration

In addition, DocXOptions has a new property, styleConfiguration, which holds an optional DocXStyleConfiguration.

* Add new initializers for DocXStylesConfiguration

The designated initializer now takes an AEXMLDocument. And there are now two new convenience initializers, one that takes a String? and one that takes a URL?.

* additional tests for styles

Co-authored-by: Brad Andalman <bya@180g.co>
Co-authored-by: Morten Bertz <morten@telethon.jp>