Fixed underlining when no explicit foreground color is set (#37)
* Underlining doesn’t require specifying a foreground color

Made `color` optional when creating an underline element. In addition, if the `underlineColor` attribute is present, it will be preferred over `foregroundColor` (if any). If neither is specified, the underline color will not be output explicitly.

* No longer explicitly look for `foregroundColor` when setting color on an underline

When no underline color is explicitly specified on the <w:u> element, Word already falls back to the text color. So, there’s no reason for DocX to perform this check as well.