Oneeightyg issue41 (#45)
* Exported images use an extension that matches their format

This fixes issue #41 (https://github.com/shinjukunian/DocX/issues/41).

Now, instead of always using the “png” extension for exported images, DocX looks at the NSTextAttachment’s fileType (UTI) and, if it finds a supported one, uses the appropriate extension.

Currently, the supported image types are: GIF, JPEG, PNG, TIFF, PDF, PSD. While there are others that are supported by Word, they are either rare or Windows-related. Since this is a Mac library, we opted not to support the latter.

Lastly, there are Mac-supported image file types (like HEIC) that aren’t supported by Word. When DocX finds an image with an unsupported type, it will attempt to convert it to a PNG. If it can, then it will export that image to docx.

* fixed compilation for iOS
added tests

* use different FileManager API to deploy test to macOS11

---------

Co-authored-by: Brad Andalman <bya@180g.co>
Co-authored-by: Morten Bertz <morten@telethon.jp>